LPD-47713 - Centralize our js helpers that replaced bootstrap components and put behind a deprecated feature flag
