+++ modules/_node-scripts/preflight/checkPackageJSONFiles.mjs
@@ -271,12 +271,9 @@ const ALLOWED_NAMED_SCOPE_EXCEPTIONS = [
 	'frontend-image-editor-capability-rotate',
 	'frontend-image-editor-capability-saturation',
 	'frontend-image-editor-web',
-	'frontend-js-alert-support-web',
 	'frontend-js-aui-web',
 	'frontend-js-clay-sample-web',
-	'frontend-js-collapse-support-web',
 	'frontend-js-components-web',
-	'frontend-js-dropdown-support-web',
 	'frontend-js-jquery-web',
 	'frontend-js-loader-modules-extender',
 	'frontend-js-lodash-web',
