+++ modules/apps/frontend-js/frontend-js-bootstrap-support-web/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Frontend JS Bootstrap Support Web
+Bundle-SymbolicName: com.liferay.frontend.js.bootstrap.support.web
+Bundle-Version: 1.0.0
+Web-ContextPath: /frontend-js-bootstrap-support-web
