+++ modules/apps/frontend-js/frontend-js-bootstrap-support-web/node-scripts.config.js
