+++ modules/apps/frontend-js/frontend-js-bootstrap-support-web/package.json
@@ -1,0 +1,15 @@
+{
+	"dependencies": {
+		"dom-align": "1.12.2",
+		"frontend-js-web": "*"
+	},
+	"main": "index.ts",
+	"name": "@liferay/frontend-js-bootstrap-support-web",
+	"private": true,
+	"scripts": {
+		"build": "node-scripts build",
+		"checkFormat": "node-scripts format --check",
+		"format": "node-scripts format"
+	},
+	"version": "1.0.0"
+}
