+++ modules/apps/frontend-js/frontend-js-bootstrap-support-web/src/main/resources/META-INF/resources/CollapseProvider.ts
