--- modules/apps/frontend-js/frontend-js-collapse-support-web/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Frontend JS Collapse Support Web
-Bundle-SymbolicName: com.liferay.frontend.js.collapse.support.web
-Bundle-Version: 2.0.27
-Web-ContextPath: /frontend-js-collapse-support-web
