--- modules/apps/frontend-js/frontend-js-collapse-support-web/package.json
@@ -1,14 +1,0 @@
-{
-	"dependencies": {
-		"frontend-js-web": "*"
-	},
-	"main": "index.ts",
-	"name": "frontend-js-collapse-support-web",
-	"private": true,
-	"scripts": {
-		"build": "node-scripts build",
-		"checkFormat": "node-scripts format --check",
-		"format": "node-scripts format"
-	},
-	"version": "2.0.27"
-}
