--- modules/apps/frontend-js/frontend-js-collapse-support-web/src/main/java/com/liferay/frontend/js/collapse/support/web/internal/servlet/taglib/CollapseBottomDynamicInclude.java
@@ -1,65 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.frontend.js.collapse.support.web.internal.servlet.taglib;
-
-import com.liferay.frontend.js.loader.modules.extender.esm.ESImportUtil;
-import com.liferay.portal.kernel.servlet.taglib.DynamicInclude;
-import com.liferay.portal.kernel.servlet.taglib.aui.JSFragment;
-import com.liferay.portal.kernel.servlet.taglib.aui.ScriptData;
-import com.liferay.portal.url.builder.AbsolutePortalURLBuilder;
-import com.liferay.portal.url.builder.AbsolutePortalURLBuilderFactory;
-
-import java.io.IOException;
-
-import java.util.Arrays;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Marko Cikos
- */
-@Component(service = DynamicInclude.class)
-public class CollapseBottomDynamicInclude implements DynamicInclude {
-
-	@Override
-	public void include(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse, String key)
-		throws IOException {
-
-		ScriptData scriptData = new ScriptData();
-
-		AbsolutePortalURLBuilder absolutePortalURLBuilder =
-			_absolutePortalURLBuilderFactory.getAbsolutePortalURLBuilder(
-				httpServletRequest);
-
-		scriptData.append(
-			null,
-			new JSFragment(
-				"main();",
-				Arrays.asList(
-					ESImportUtil.getESImport(
-						absolutePortalURLBuilder,
-						"{main} from frontend-js-collapse-support-web"))));
-
-		scriptData.writeTo(httpServletResponse.getWriter());
-	}
-
-	@Override
-	public void register(
-		DynamicInclude.DynamicIncludeRegistry dynamicIncludeRegistry) {
-
-		dynamicIncludeRegistry.register("/html/common/themes/bottom.jsp#post");
-	}
-
-	@Reference
-	private AbsolutePortalURLBuilderFactory _absolutePortalURLBuilderFactory;
-
-}
