--- modules/apps/frontend-js/frontend-js-collapse-support-web/src/main/resources/META-INF/resources/index.ts
@@ -1,10 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import CollapseProvider from './CollapseProvider';
-
-export function main() {
-	new CollapseProvider();
-}
