--- modules/apps/frontend-js/frontend-js-collapse-support-web/src/main/resources/META-INF/resources/tsconfig.json
@@ -1,39 +1,0 @@
-{
-	"@generated": "337dcbef372bdac9994834df987d3b6f7edff6fe",
-	"@readonly": "** AUTO-GENERATED: DO NOT EDIT **",
-	"compilerOptions": {
-		"allowSyntheticDefaultImports": true,
-		"baseUrl": ".",
-		"checkJs": false,
-		"composite": true,
-		"declarationDir": "../../../../../../../../.tsc/types",
-		"emitDeclarationOnly": true,
-		"jsx": "react",
-		"module": "es6",
-		"moduleResolution": "node",
-		"paths": {
-			"frontend-js-web": [
-				"../../../../../../frontend-js-web/src/main/resources/META-INF/resources/index.d.ts"
-			]
-		},
-		"rootDir": "../../../../../../../..",
-		"skipLibCheck": true,
-		"sourceMap": false,
-		"strict": true,
-		"target": "esnext",
-		"tsBuildInfoFile": "../../../../../../../../.tsc/buildinfo/frontend-js-collapse-support-web.tsbuildinfo",
-		"typeRoots": [
-			"../../../../../../../../node_modules/@types"
-		]
-	},
-	"include": [
-		"**/*.ts",
-		"**/*.tsx",
-		"../../../../../../../../global.d.ts"
-	],
-	"references": [
-		{
-			"path": "../../../../../../frontend-js-web/src/main/resources/META-INF/resources/tsconfig.json"
-		}
-	]
-}
