--- modules/apps/frontend-js/frontend-js-dropdown-support-web/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Frontend JS Dropdown Support Web
-Bundle-SymbolicName: com.liferay.frontend.js.dropdown.support.web
-Bundle-Version: 2.0.24
-Web-ContextPath: /frontend-js-dropdown-support-web
