--- modules/apps/frontend-js/frontend-js-dropdown-support-web/build.gradle
@@ -1,9 +1,0 @@
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
-	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.4.0"
-	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
-	compileOnly project(":apps:frontend-js:frontend-js-loader-modules-extender-api")
-	compileOnly project(":apps:portal-url-builder:portal-url-builder-api")
-	compileOnly project(":core:petra:petra-string")
-}
