--- modules/apps/frontend-js/frontend-js-dropdown-support-web/node-scripts.config.js
@@ -1,8 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-module.exports = {
-	main: './src/main/resources/META-INF/resources/index.ts',
-};
