--- modules/apps/frontend-js/frontend-js-dropdown-support-web/package.json
@@ -1,15 +1,0 @@
-{
-	"dependencies": {
-		"dom-align": "1.12.2",
-		"frontend-js-web": "*"
-	},
-	"main": "index.ts",
-	"name": "frontend-js-dropdown-support-web",
-	"private": true,
-	"scripts": {
-		"build": "node-scripts build",
-		"checkFormat": "node-scripts format --check",
-		"format": "node-scripts format"
-	},
-	"version": "2.0.24"
-}
