--- modules/apps/frontend-js/frontend-js-dropdown-support-web/src/main/resources/META-INF/resources/tsconfig.json
@@ -1,39 +1,0 @@
-{
-	"@generated": "a37ca0a75cfc51d7749251de8ae8a73c3cbf017e",
-	"@readonly": "** AUTO-GENERATED: DO NOT EDIT **",
-	"compilerOptions": {
-		"allowSyntheticDefaultImports": true,
-		"baseUrl": ".",
-		"checkJs": false,
-		"composite": true,
-		"declarationDir": "../../../../../../../../.tsc/types",
-		"emitDeclarationOnly": true,
-		"jsx": "react",
-		"module": "es6",
-		"moduleResolution": "node",
-		"paths": {
-			"frontend-js-web": [
-				"../../../../../../frontend-js-web/src/main/resources/META-INF/resources/index.d.ts"
-			]
-		},
-		"rootDir": "../../../../../../../..",
-		"skipLibCheck": true,
-		"sourceMap": false,
-		"strict": true,
-		"target": "esnext",
-		"tsBuildInfoFile": "../../../../../../../../.tsc/buildinfo/frontend-js-dropdown-support-web.tsbuildinfo",
-		"typeRoots": [
-			"../../../../../../../../node_modules/@types"
-		]
-	},
-	"include": [
-		"**/*.ts",
-		"**/*.tsx",
-		"../../../../../../../../global.d.ts"
-	],
-	"references": [
-		{
-			"path": "../../../../../../frontend-js-web/src/main/resources/META-INF/resources/tsconfig.json"
-		}
-	]
-}
