--- modules/apps/frontend-js/frontend-js-tabs-support-web/package.json
@@ -1,14 +1,0 @@
-{
-	"dependencies": {
-		"frontend-js-web": "*"
-	},
-	"main": "index.ts",
-	"name": "frontend-js-tabs-support-web",
-	"private": true,
-	"scripts": {
-		"build": "node-scripts build",
-		"checkFormat": "node-scripts format --check",
-		"format": "node-scripts format"
-	},
-	"version": "2.0.25"
-}
