--- modules/apps/frontend-js/frontend-js-tabs-support-web/src/main/resources/META-INF/resources/index.ts
@@ -1,10 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import TabsProvider from './TabsProvider';
-
-export function main() {
-	new TabsProvider();
-}
