--- modules/apps/frontend-js/frontend-js-tabs-support-web/src/main/resources/META-INF/resources/tsconfig.json
@@ -1,39 +1,0 @@
-{
-	"@generated": "613fe9aa5c140e0ed11c83dacb4330bed7b94dc4",
-	"@readonly": "** AUTO-GENERATED: DO NOT EDIT **",
-	"compilerOptions": {
-		"allowSyntheticDefaultImports": true,
-		"baseUrl": ".",
-		"checkJs": false,
-		"composite": true,
-		"declarationDir": "../../../../../../../../.tsc/types",
-		"emitDeclarationOnly": true,
-		"jsx": "react",
-		"module": "es6",
-		"moduleResolution": "node",
-		"paths": {
-			"frontend-js-web": [
-				"../../../../../../frontend-js-web/src/main/resources/META-INF/resources/index.d.ts"
-			]
-		},
-		"rootDir": "../../../../../../../..",
-		"skipLibCheck": true,
-		"sourceMap": false,
-		"strict": true,
-		"target": "esnext",
-		"tsBuildInfoFile": "../../../../../../../../.tsc/buildinfo/frontend-js-tabs-support-web.tsbuildinfo",
-		"typeRoots": [
-			"../../../../../../../../node_modules/@types"
-		]
-	},
-	"include": [
-		"**/*.ts",
-		"**/*.tsx",
-		"../../../../../../../../global.d.ts"
-	],
-	"references": [
-		{
-			"path": "../../../../../../frontend-js-web/src/main/resources/META-INF/resources/tsconfig.json"
-		}
-	]
-}
