+++ modules/apps/portal-language/portal-language-lang/src/main/resources/content/Language.properties
@@ -7584,6 +7584,8 @@ feature.flag.LPD-44091.description=Walkthrough is deprecated and will be removed
 feature.flag.LPD-44091.title=Walkthrough
 feature.flag.LPD-46393.description=Dynamically show or hide configuration fields in a custom fragment based on user input values.
 feature.flag.LPD-46393.title=Conditional Configuration Field Visibility in Custom Fragments
+feature.flag.LPD-47713.description=The support for global Javascript Tabs component provided by Liferay is deprecated.
+feature.flag.LPD-47713.title=JS Component Support: Tabs
 feature.flag.LPS-96845.description=Configured SCIM Clients can leverage REST APIs to sync identity information for users and groups with external applications.
 feature.flag.LPS-96845.title=System for Cross-domain Identity Management (SCIM)
 feature.flag.LPS-122920.description=Create a search experience that understands the meaning and the context of your content and users searches leveraging language ML models through third-party text embedding providers and Search Blueprints. Once enabled, begin configuring in Instance Settings > Search Experiences > Semantic Search.
