+++ portal-impl/src/portal.properties
@@ -6694,6 +6694,11 @@
     feature.flag.LPD-46393=false
 
     #
+    # Env: LIFERAY_FEATURE_PERIOD_FLAG_PERIOD__UPPERCASEL__UPPERCASEP__UPPERCASED__MINUS__NUMBER4__NUMBER7__NUMBER7__NUMBER1__NUMBER3_
+    #
+    feature.flag.LPD-47713=true
+
+    #
     # Env: LIFERAY_FEATURE_PERIOD_FLAG_PERIOD__UPPERCASEL__UPPERCASEP__UPPERCASES__MINUS__NUMBER1__NUMBER2__NUMBER2__NUMBER9__NUMBER2__NUMBER0_
     #
     feature.flag.LPS-122920=false
@@ -7101,6 +7106,11 @@
     # Env: LIFERAY_FEATURE_PERIOD_FLAG_PERIOD__UPPERCASEL__UPPERCASEP__UPPERCASED__MINUS__NUMBER4__NUMBER6__NUMBER3__NUMBER9__NUMBER3__PERIOD_TYPE
     #
     feature.flag.LPD-46393.type=release
+
+    #
+    # Env: LIFERAY_FEATURE_PERIOD_FLAG_PERIOD__UPPERCASEL__UPPERCASEP__UPPERCASED__MINUS__NUMBER4__NUMBER7__NUMBER7__NUMBER1__NUMBER3__PERIOD_TYPE
+    #
+    feature.flag.LPD-47713.type=deprecation
 
     #
     # Env: LIFERAY_FEATURE_PERIOD_FLAG_PERIOD__UPPERCASEL__UPPERCASEP__UPPERCASES__MINUS__NUMBER1__NUMBER2__NUMBER2__NUMBER9__NUMBER2__NUMBER0__PERIOD_TYPE
